+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ca.dart
@@ -997,6 +997,18 @@ class _AccessibilityTranslationsCa extends AccessibilityTranslations {
   }
 
   @override
+  String outgoingMessageLabel({required String body}) => 'Has dit, $body';
+
+  @override
+  String incomingMessageLabel({required String senderName, required String body}) => '$senderName ha dit, $body';
+
+  @override
+  String outgoingDeletedMessageLabel({required String body}) => 'Tu, $body';
+
+  @override
+  String incomingDeletedMessageLabel({required String senderName, required String body}) => '$senderName, $body';
+
+  @override
   String get pollPreviewLabel => 'Enquesta';
 
   @override