+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_de.dart
@@ -995,6 +995,18 @@ class _AccessibilityTranslationsDe extends AccessibilityTranslations {
   }
 
   @override
+  String outgoingMessageLabel({required String body}) => 'Du hast gesagt, $body';
+
+  @override
+  String incomingMessageLabel({required String senderName, required String body}) => '$senderName hat gesagt, $body';
+
+  @override
+  String outgoingDeletedMessageLabel({required String body}) => 'Du, $body';
+
+  @override
+  String incomingDeletedMessageLabel({required String senderName, required String body}) => '$senderName, $body';
+
+  @override
   String get pollPreviewLabel => 'Umfrage';
 
   @override