+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_no.dart
@@ -980,6 +980,18 @@ class _AccessibilityTranslationsNo extends AccessibilityTranslations {
   }
 
   @override
+  String outgoingMessageLabel({required String body}) => 'Du sa, $body';
+
+  @override
+  String incomingMessageLabel({required String senderName, required String body}) => '$senderName sa, $body';
+
+  @override
+  String outgoingDeletedMessageLabel({required String body}) => 'Du, $body';
+
+  @override
+  String incomingDeletedMessageLabel({required String senderName, required String body}) => '$senderName, $body';
+
+  @override
   String get pollPreviewLabel => 'Avstemning';
 
   @override